Improve the UX of the admin page
## What is this?
This improves the ux of the admin page

## Why this?
The current admin page does not have great experience, user need to
navigate multiple times to add images and updates. Also, you don't have
a great view of things on the dashboard. This improves the UX greatly
for smoother experience

## How are you doing this?
I'm leveraging on existing django-admin tools